// src/app/modules/ci/model/globales_CI.interface.ts
export interface Iglobales_CI {
	CostosDec: number;
	ExistenciasDec: number;
	PesosDec: number;
	costoFiscal: string;
	costoCorporativo: string;
	costoIngresoEspecial: string;
	UnidadPeso: string;
	UnidadVolumen: string;
	cuarentena: boolean;
	disponible: boolean;
	reserva: boolean;
	vencida: boolean;
	NombreClasif1: string;
	NombreClasif2: string;
	NombreClasif3: string;
	NombreClasif4: string;
	NombreClasif5: string;
	NombreClasif6: string;
	IntegracionConta: boolean;
	UsaLocalizacion: boolean;
	UsaCodigoBarras: boolean;
	UsarAprobacion: boolean;
	relArticuloProveedor: boolean;
	resExistenciaDocumento: boolean;
	kits: boolean;
	redondearUnidadDetalle: boolean;
	TransacXUsuario: boolean;
	UsaConsecutivos: boolean;
	manejaEstandar: boolean;
	UsarNumerosSerie: boolean;
	LineasMaxTrans: number;
	TipoAsientoId: number;
	PaqueteContableId: number;
	generacionAsientoVisual: boolean;
	generacionAsientoModificar: boolean;
	generacionAsientoAplicar: boolean;
	AsntAjuCompra: boolean;
	AsntAjuConsumo: boolean;
	AsntAjuCosto: boolean;
	AsntAjuFisico: boolean;
	AsntAjuMiscelan: boolean;
	AsntAjuProduc: boolean;
	AsntAjuVencim: boolean;
	AsntAjuVenta: boolean;
	CtrEnTransaccion: boolean;
	UsaUnidadesDist: boolean;
	AsistenciaAutomat: boolean;
	UsaCodigoEAN13: boolean;
	EAN13ReglaLocal: string;
	UsaCodigoEAN8: boolean;
	UsaCodigoUCC12: boolean;
	UCC12ReglaLocal: string;
	UsaCodigoUCC8: boolean;
	EAN8ReglaLocal: string;
	UsaCodigoGeneric: boolean;
	ModalidadUso: string;
	PrioridadBusqueda: string;
	AjustarConteo: boolean;
	ExistEnTotales: string;
	FchUltProcAprob: string;
	FchUltProcVcto: string;
	FechaInicioTrans: string;
	ModAplicAsiento: number;
	CntrlSeriesEntr: boolean;
}