// src/app/modules/ci/administracion/parametros-modulo/service/parametros-ci-api-model-interface.ts
export interface ICreateGlobalesCI {
	id: number;
	costoDecimal: number;
	existenciaDecimal: number;
	pesoDecimal: number;
	unidadPeso: string;
	unidadVolumen: string;
	usaLocalizaciones: boolean;
	AjustarConteo: boolean;
	maxAuditoria: number;
	isDeleted: boolean;
	createBy: string;
	createDate: string;
	updateBy: string;
	updateDate: string;
	asntAjuCompras: boolean;
	asntAjuConsumo: boolean;
	asntAjuCosto: boolean;
	astnAjuFisico: boolean;
	asntAjuMiscelan: boolean;
	asntAjuProduc: boolean;
	asntAjuVencim: boolean;
	asntAjuVent: boolean;
	ctrEnTransaccion: boolean;
	existEnTotales: string;
	fchUltProcApro: string;
	fchUltProcVcto: string;
	fechaInicioTrans: string;
	integracionCont: boolean;
	modAplicAsiento: boolean;
	nombreClasif1: string;
	nombreClasif2: string;
	nombreClasif3: string;
	nombreClasif4: string;
	nombreClasif5: string;
	nombreClasif6: string;
	paqueteId: number;
	tipoAsientoId: number;
	asistenciaAutomatica: boolean;
	cntrlSeriesEntr: boolean;
	EAN13ReglaLocal: string;
	EAN18ReglaLocal: string;
	lienaMaxTrans: number;
	modalidadUso: string;
	prioridadBusqueda: string;
	transacXusuario: boolean;
	UCC12ReglaLocal: string;
	usaCodigoBarras: boolean;
	usaCodigoEAN13: boolean;
	usaCodigoEAN8: boolean;
	usaCodigoGeneric: boolean;
	usaCodigoUCC12: boolean;
	usaCodigoUCC8: boolean;
	usaConsecutivos: boolean;
	usaUnidadDistr: boolean;
	usarAprobacion: boolean;
	usarNumeroSerie: boolean;
}